Rename orbit station agent to station agent, and reduce files where it is used
